Create README.md
I am ETHO MBA Steeve Rayanne, from Gabon as a 2nd year student of software engineering. First of all, my studies allow me to acquire knowledge and practice in the field of computer science. Indeed, thanks to the courses accumulated executed during these semesters, I can easily describe the functionalities of a computer or create a database of any company.
For this course, I would like to acquire knowledge advantages, learn and soak up this information but above all validate it to be able to advance in my study program.
For the rest, I was lucky to have learned to drive from my parents, which means that sometimes I have the opportunity to drive here thanks to my license but I don’t do it all the time, unless it’s to visit a province or whatever.